优化high和low模型在运行时的load和offload速度，解决多卡偶发卡住问题，优化lora加载，去除多卡中不必要的类型转换 (merge request !67)
Squash merge branch 'gpu' into 'main'
优化high和low模型在运行时的load和offload速度
解决多卡偶发卡住问题
优化lora加载
去除多卡中不必要的类型转换
2卡
<img width="" src="/uploads/D36F4145122646538DEA65CB9591F2FC/image.png" alt="image.png" />

+++ ksana/distributed/sequence_parallel.py
@@ -51,7 +51,7 @@ def rope_apply(x, grid_sizes, freqs):
 
         # append to collection
         output.append(x_i)
-    return torch.stack(output).float()
+    return torch.stack(output).type_as(x)
 
 
 def sp_attn_forward(self, x, seq_lens, grid_sizes, freqs):
